1.8.0 fixup (#159)
* removed datasource

* moved urlsession conformance to core

* fixes

* headers

* lifecycle migration

* podspec

* background sleep

* updated method names to include 'service' instead of 'profile' (#157)

* wip

* WIP

* wip

* wip

* Remove UIWebView (#158)

* update settings

* remove uiwv start

* remove uiwv start 2

* updated avail checks

* Clean up removing UIWebView

* Change teal helper back to tealiummobile/demo

* wip

* wip

* fixed dispatchqueue test

* tests

* cartfile

* removed .framework files

* removed comments

+++ tealium/core/CoreUtils/TealiumExtensions.swift
@@ -89,8 +89,8 @@ extension Date {
         return String(describing: time)
     }
 
-    public func millisecondsFrom(earlierDate: Date) -> Int {
-        return Int(self.timeIntervalSince(earlierDate) * 1000)
+    public func millisecondsFrom(earlierDate: Date) -> Int64 {
+        return Int64(self.timeIntervalSince(earlierDate) * 1000)
     }
 }
